+++ src/plugins/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Plugins
+ *
+ * Re-exports all plugin modules.
+ */
+
+export * from './renderers';
+export * from './primitives';