+++ src/plugins/renderers/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Renderer Plugins
+ *
+ * Re-exports all render pipeline plugins.
+ */
+
+export { LineRenderer } from './line';